Update DWIN.h
DWIN_HMI_icon_display() update

+++ src/DWIN.h
@@ -121,7 +121,7 @@ public:
   bool setText_crc(long address, String textData);
 
   // Icon Display (from 48.ICL file library) on VP Address of 'Basic Graphic' item (send order using CRC)
-  bool iconDisplay_crc(long address, int lib_icon, int x, int y, int icon);
+  bool iconDisplay_crc(long address, uint8_t lib_icon, int x, int y, int icon);
 
   // Set DGUS_SERIAL UART baudrate 9600
   void setDsBaudrate_9600();
